backend/src/requirements.txt
asgiref==3.5.2
Django==3.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djongo==1.3.6
Markdown==3.4.1
pymongo==3.12.1
pytz==2022.2.1
sqlparse==0.2.4
tzdata==2022.2
